% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myconstr.R
\name{myconstr}
\alias{myconstr}
\title{Constructor Function}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{x}{named list}

\item{y}{named list}

\item{alpha}{confidence inveral}
}
\value{
a stats_list of my constructor function
}
\description{
Constructor Function
}
