% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Boxplot}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{add from the ggplot function}
}
\value{
a plot of Rttest
}
\description{
Boxplot
}
