% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Print Rttest function}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{named list}

\item{...}{extra}
}
\value{
a kable of the columns x and y
}
\description{
Print Rttest function
}
