% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\name{myconstr}
\alias{myconstr}
\title{Constructor Function}
\usage{
myconstr(x, y, alpha)
}
\arguments{
\item{alpha}{}
}
\value{
a stats_list of my constructor function
}
\description{
Constructor Function
}
